+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: allefebv <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/09 10:05:25 by allefebv          #+#    #+#             */
-/*   Updated: 2018/11/15 12:28:42 by allefebv         ###   ########.fr       */
+/*   Updated: 2018/11/15 14:04:50 by allefebv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_strlen(const char *s)
+#include <string.h>
+
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
