+++ ft_strrev.c
@@ -6,7 +6,7 @@
 /*   By: allefebv <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/12 21:21:53 by allefebv          #+#    #+#             */
-/*   Updated: 2018/11/15 14:01:15 by allefebv         ###   ########.fr       */
+/*   Updated: 2018/11/15 14:14:26 by allefebv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strrev(char *s)
 {
-	int		start;
-	int		end;
+	size_t	start;
+	size_t	end;
 	char	tmp;
 
 	start = 0;
